Step 11 - Enemy 4 and Final Touches
Enemy 4 is now scripted to drift around the map and is pretty difficult
to kill with the standard blaster as it has an array of parts which all
have health that needs to be taken down.

Additionally, a space background has been added that is scripted to move
along two transparent fields to make simulate moving.

At this point, Gibson doesn't really mention any more to be done, but
I'm going to try and get around to make the score text and lives.